DOESN'T CLOSE: #670 Add URL import functionality and confirmation prompts for settings import
Add the ability to share a url to import settings. Also has a confirm prompt to make sure nothing gets overwritten
IT DOES NOT WORK ON FIREFOX, the extension guid is always different. Not tested on chrome.
SO IT DOESN'T CLOSE: [Feature Request] Load JSON from web address as workaround to cookies. #670